requirements.txt
fastapi
httpx
jinja2
locust
pillow
psycopg[binary]
python-multipart
uvicorn[standard]